requirements.txt
langchain>=0.2.14
langchain-core>=0.2.33
langchain-community>=0.2.12
langchain-openai>=0.1.22
langchain-mistralai>=0.1.12
langgraph>=0.2.14
langchain-neo4j>=0.0.1
asyncpg>=0.29.0
pgvector>=0.3.0
pydantic>=2.8.2
fastapi>=0.112.1
uvicorn>=0.30.6
python-dotenv>=1.0.1
ragas>=0.1.14
sentence-transformers>=3.0.1
